+++ source/Detector.cpp
@@ -2265,7 +2265,6 @@ void Detector::applyOpenShutterSmearing(float exposureTime)
             if(std::isnan(rowFOV))
             {
                 // All exposed rows are within the FOV
-
                 numExposedRowsInFOV(column) = numRows - firstRowExposed;
             }
             else
@@ -2376,14 +2375,32 @@ void Detector::applyOpenShutterSmearing(float exposureTime)
  **/
 double Detector::getRowEdgeFOV(int column)
 {
+    double offsetCol;
+    double offsetRow;
+    double angle;
     double pixelSizeMm = pixelSize / 1000.0;    // Pixel size [µm] -> [mm]
 
+    if (ccdPosition == "Custom")
+    {
+        offsetCol = customOriginOffsetX;
+        offsetRow = customOriginOffsetY;
+        angle   = customOrientationAngle;
+    }
+    else
+    {
+        int ccd = stoi(ccdPosition) - 1;
+        array<double, 12> currentCcdPositions = (*ccdPositions)();
+        offsetCol = currentCcdPositions[ccd * 3];
+        offsetRow = currentCcdPositions[ccd * 3 + 1];
+        angle   = deg2rad(currentCcdPositions[ccd * 3 + 2]);
+    }
+  
     // Quadratic equation: a * x**2 + b * x + c  = 0
     // Find intersection between circle representing the FOV and the given column
 
     double a = pow(pixelSizeMm, 2);
-    double b = - 2 * pixelSizeMm * subFieldZeroPointRow;
-    double c = pow(subFieldZeroPointRow, 2) + pow(column * pixelSizeMm - subFieldZeroPointColumn , 2) -  pow(camera.getFocalLength() * tan(radiusFOV), 2);
+    double b = 2 * pixelSizeMm * ( pixelSizeMm * subFieldZeroPointRow - offsetRow);
+    double c = pow(pixelSizeMm * subFieldZeroPointRow - offsetRow, 2) + pow((column + subFieldZeroPointColumn) * pixelSizeMm - offsetCol , 2) -  pow(camera.getFocalLength() * tan(radiusFOV), 2);
 
     // Discriminant (should be positive)
 
@@ -2592,7 +2609,7 @@ void Detector::applyGain()
     combinedGainLeft = frontEndElectronics->getGainLeftAdc(internalTime) * ccdGainLeft;
     combinedGainRight = frontEndElectronics->getGainRightAdc(internalTime) * ccdGainRight;
 
-    if(lastIndexSubFieldLeft >= numColumnsPixelMap - 1)      // Left ADC only
+    if(lastIndexSubFieldLeft >= ((int) numColumnsPixelMap - 1))      // Left ADC only
     {
         pixelMap *= combinedGainLeft;
         smearingMap *= combinedGainLeft;