+++ CHANGELOG.md
@@ -8,8 +8,12 @@
 ## **/**/****: 3.5.4
 
 ## Fixed
+* Fixed bug where incorrect gain was applied to the right CCD.
+* Fixed a bug when applying open shutter smearing when we include relative transmissivity. 
 
 ## Changed
+
+* Changed the test for the gain, to correctly test that the left/right CCD have different gains.
 
 ## Added